order and payment implementation

+++ FoodDelivery/restaurantListingApp/models.py
@@ -6,5 +6,5 @@ class Restaurants(models.Model):
     id = models.BigAutoField(primary_key=True, editable=False)
     name = models.CharField(max_length=158, null=False, blank=False)
     location = models.CharField(max_length=500, null=False, blank=False)
-    rating = models.IntegerField(null=True, blank=True)
+    rating = models.FloatField(null=True, blank=True)
